refactor(mobile): share the stocks pill grid
Extract a reusable StockPillRows component (2-row scrollable MediaButton grid +
skeleton) into mvvm/components, and rewire both the Global Search StocksSection
and the Portfolio StocksDiscoverySection to use it, removing the duplicated
grid/pill markup. Finalize stocks subheader spacing: holdings list s4 (like the
other sections), discovery grid s12.

LIVE-27854
