+++ .changeset/shared-stock-pill-rows.md
@@ -1,0 +1,5 @@
+---
+"live-mobile": patch
+---
+
+Extract a shared `StockPillRows` component (two horizontally-scrollable rows of stock pills) used by both Global Search and the Portfolio Stocks discovery section, removing the duplicated grid. Finalize the Stocks subheader spacing: the holdings list matches the other portfolio sections, the discovery grid uses s12.
