+++ apps/ledger-live-mobile/src/mvvm/features/GlobalSearch/screens/GlobalSearch/components/StocksSection/index.tsx
@@ -1,17 +1,14 @@
-import React, { useCallback, useMemo } from "react";
-import { ScrollView } from "react-native";
-import Icon from "@ledgerhq/crypto-icons/native";
+import React, { useCallback } from "react";
 import type { StockSuggestion } from "@ledgerhq/live-common/dada-client/utils/assetDiscovery";
 import {
   Box,
-  MediaButton,
-  Skeleton,
   Subheader,
   SubheaderRow,
   SubheaderShowMore,
   SubheaderTitle,
 } from "@ledgerhq/lumen-ui-rnative";
 import type { LumenViewStyle } from "@ledgerhq/lumen-ui-rnative/styles";
+import { StockPillRows } from "LLM/components/StockPillRows";
 import type { GlobalSearchCategory } from "../../useGlobalSearchViewModel";
 
 const SKELETON_PILLS_PER_ROW = 4;
@@ -26,36 +23,6 @@ type Props = Readonly<{
   onStockPress: (stock: StockSuggestion) => void;
 }>;
 
-function splitIntoRows(stocks: StockSuggestion[]): [StockSuggestion[], StockSuggestion[]] {
-  const top: StockSuggestion[] = [];
-  const bottom: StockSuggestion[] = [];
-  stocks.forEach((stock, index) => (index % 2 === 0 ? top : bottom).push(stock));
-  return [top, bottom];
-}
-
-type StockPillProps = Readonly<{
-  stock: StockSuggestion;
-  onPress: (stock: StockSuggestion) => void;
-}>;
-
-function StockPill({ stock, onPress }: StockPillProps) {
-  const handlePress = useCallback(() => onPress(stock), [stock, onPress]);
-
-  return (
-    <MediaButton
-      size="sm"
-      hideChevron
-      leadingContentShape="rounded"
-      leadingContent={<Icon ledgerId={stock.ledgerId} ticker={stock.ticker} size={24} />}
-      onPress={handlePress}
-      accessibilityLabel={stock.name}
-      testID={`global-search-stock-${stock.ticker.toLowerCase()}`}
-    >
-      {stock.ticker.toUpperCase()}
-    </MediaButton>
-  );
-}
-
 export function StocksSection({
   title,
   testID,
@@ -65,15 +32,9 @@ export function StocksSection({
   onSeeAll,
   onStockPress,
 }: Props) {
-  const [topRow, bottomRow] = useMemo(() => splitIntoRows(stocks), [stocks]);
   const handleSeeAll = useCallback(() => onSeeAll(category), [onSeeAll, category]);
 
   if (!isLoading && stocks.length === 0) return null;
-
-  const renderSkeletonPills = () =>
-    Array.from({ length: SKELETON_PILLS_PER_ROW }, (_, index) => (
-      <Skeleton key={index} lx={pillSkeletonStyle} />
-    ));
 
   return (
     <Box lx={sectionStyle} testID={testID}>
@@ -89,40 +50,16 @@ export function StocksSection({
           </SubheaderRow>
         </Subheader>
       </Box>
-      {isLoading ? (
-        <Box lx={insetStyle}>
-          <Box lx={skeletonGridStyle}>
-            <Box lx={rowStyle}>{renderSkeletonPills()}</Box>
-            <Box lx={rowStyle}>{renderSkeletonPills()}</Box>
-          </Box>
-        </Box>
-      ) : (
-        <ScrollView
-          horizontal
-          showsHorizontalScrollIndicator={false}
-          keyboardShouldPersistTaps="handled"
-        >
-          <Box lx={gridStyle}>
-            <Box lx={rowStyle}>
-              {topRow.map(stock => (
-                <StockPill key={stock.id} stock={stock} onPress={onStockPress} />
-              ))}
-            </Box>
-            <Box lx={rowStyle}>
-              {bottomRow.map(stock => (
-                <StockPill key={stock.id} stock={stock} onPress={onStockPress} />
-              ))}
-            </Box>
-          </Box>
-        </ScrollView>
-      )}
+      <StockPillRows
+        stocks={stocks}
+        isLoading={isLoading}
+        onStockPress={onStockPress}
+        skeletonPillsPerRow={SKELETON_PILLS_PER_ROW}
+        testIdPrefix="global-search-stock"
+      />
     </Box>
   );
 }
 
 const sectionStyle: LumenViewStyle = { gap: "s16", marginHorizontal: "-s16" };
 const insetStyle: LumenViewStyle = { paddingHorizontal: "s16" };
-const gridStyle: LumenViewStyle = { gap: "s8", paddingHorizontal: "s16" };
-const skeletonGridStyle: LumenViewStyle = { gap: "s8" };
-const rowStyle: LumenViewStyle = { flexDirection: "row", gap: "s8" };
-const pillSkeletonStyle: LumenViewStyle = { width: "s96", height: "s40", borderRadius: "full" };
